+++ src/lib/dglib/include/DgDistance.h
@@ -30,7 +30,7 @@ class DgDistanceBase {
 
       long double asDouble (void) const { return rf().toDouble(*this); }
 
-      unsigned long long int asInt (void) const { return rf().toInt(*this); }
+      uint64_t asInt (void) const { return rf().toInt(*this); }
 
    protected:
